+++ tests/image/classification/test_data.py
@@ -167,8 +167,8 @@ def test_from_filepaths_visualise(tmpdir):
 
     # call show functions
     # dm.show_train_batch()
-    dm.show_train_batch("pre_tensor_transform")
-    dm.show_train_batch(["pre_tensor_transform", "post_tensor_transform"])
+    dm.show_train_batch("per_sample_transform")
+    dm.show_train_batch(["per_sample_transform", "per_batch_transform"])
 
 
 @pytest.mark.skipif(not _IMAGE_AVAILABLE, reason="image libraries aren't installed.")
@@ -202,9 +202,7 @@ def test_from_filepaths_visualise_multilabel(tmpdir):
 
     # call show functions
     dm.show_train_batch()
-    dm.show_train_batch("pre_tensor_transform")
-    dm.show_train_batch("to_tensor_transform")
-    dm.show_train_batch(["pre_tensor_transform", "post_tensor_transform"])
+    dm.show_train_batch("per_sample_transform")
     dm.show_val_batch("per_batch_transform")
 
 
@@ -228,7 +226,7 @@ def test_from_filepaths_splits(tmpdir):
     assert len(train_filepaths) == len(train_labels)
 
     _to_tensor = {
-        "to_tensor_transform": nn.Sequential(
+        "per_sample_transform": nn.Sequential(
             ApplyToKeys(DataKeys.INPUT, torchvision.transforms.ToTensor()),
             ApplyToKeys(DataKeys.TARGET, torch.as_tensor),
         ),
@@ -589,7 +587,7 @@ def test_albumentations_mixup(single_target_csv):
 
     train_transform = {
         # applied only on images as ApplyToKeys is used with `input`
-        "post_tensor_transform": ApplyToKeys("input", AlbumentationsAdapter(albumentations.HorizontalFlip(p=0.5))),
+        "per_sample_transform": ApplyToKeys("input", AlbumentationsAdapter(albumentations.HorizontalFlip(p=0.5))),
         "per_batch_transform": mixup,
     }
     # merge the default transform for this task with new one.